fix(be): fixed several defects:
- incorrect db-query for a single day audits
- input request validation for get -infra-overview
- stub response for days with no data
- divide-by-zero in score calculation
