+++ src/backend/joseki.be/webapp/Handlers/InfraScoreExtensions.cs
@@ -1,0 +1,117 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+using webapp.Database;
+using webapp.Models;
+
+namespace webapp.Handlers
+{
+    /// <summary>
+    /// Common Infra-Score extensions.
+    /// </summary>
+    public static class InfraScoreExtensions
+    {
+        /// <summary>
+        /// Composes an infrastructure overview object for a requested date and provided audits.
+        /// </summary>
+        /// <param name="cache">Extended cache object.</param>
+        /// <param name="date">The date to calculate overview for.</param>
+        /// <param name="audits">List of audits to include into overview.</param>
+        /// <returns>Complete infrastructure-overview object.</returns>
+        public static async Task<InfrastructureOverview> GetInfrastructureOverview(this InfrastructureScoreCache cache, DateTime date, Database.Models.Audit[] audits)
+        {
+            // 0. if no audits - return stub
+            if (audits.Length == 0)
+            {
+                return new InfrastructureOverview
+                {
+                    Overall = new InfrastructureComponentSummaryWithHistory
+                    {
+                        Component = new InfrastructureComponent(InfrastructureScoreCache.OverallId)
+                        {
+                            Category = InfrastructureCategory.Overall,
+                            Name = "Overall infrastructure",
+                        },
+                        Date = date,
+                        ScoreHistory = new ScoreHistoryItem[0],
+                        Current = new CountersSummary(),
+                    },
+                    Components = new InfrastructureComponentSummaryWithHistory[0],
+                };
+            }
+
+            // 1. Calculate Overall data first, as it reloads caches for component audits automatically
+            var overallHistory = new List<ScoreHistoryItem>();
+            foreach (var dateTime in audits.Select(i => i.Date.Date).Distinct())
+            {
+                var historyItem = await cache.GetCountersSummary(InfrastructureScoreCache.OverallId, dateTime);
+                overallHistory.Add(new ScoreHistoryItem(dateTime, historyItem.Score));
+            }
+
+            var overall = new InfrastructureComponentSummaryWithHistory
+            {
+                Component = new InfrastructureComponent(InfrastructureScoreCache.OverallId)
+                {
+                    Category = InfrastructureCategory.Overall,
+                    Name = "Overall infrastructure",
+                },
+                Date = date,
+                ScoreHistory = overallHistory.ToArray(),
+                Current = await cache.GetCountersSummary(InfrastructureScoreCache.OverallId, date),
+            };
+
+            // 2. Calculate each component data
+            // The code expects that each date has only unique component-identifiers
+            // Thus, iterating only over requested date - gives unique components
+            var components = new List<InfrastructureComponentSummaryWithHistory>();
+            foreach (var audit in audits.Where(i => i.Date.Date == date.Date))
+            {
+                // first, find all audits for the particular component and prepare history items for it
+                var componentHistory = new List<ScoreHistoryItem>();
+                foreach (var historyAudit in audits.Where(i => i.ComponentId == audit.ComponentId))
+                {
+                    var historyItem = await cache.GetCountersSummary(historyAudit.ComponentId, historyAudit.Date);
+                    componentHistory.Add(new ScoreHistoryItem(historyAudit.Date, historyItem.Score));
+                }
+
+                // when the history is ready, calculate summary for the requested date.
+                var currentSummary = await cache.GetCountersSummary(audit.ComponentId, audit.Date);
+                var component = new InfrastructureComponentSummaryWithHistory
+                {
+                    Date = date,
+                    Component = new InfrastructureComponent(audit.ComponentId)
+                    {
+                        Category = GetCategory(audit.ComponentId),
+                        Name = audit.ComponentName,
+                    },
+                    Current = currentSummary,
+                    ScoreHistory = componentHistory.ToArray(),
+                };
+
+                components.Add(component);
+            }
+
+            return new InfrastructureOverview
+            {
+                Overall = overall,
+                Components = components.ToArray(),
+            };
+        }
+
+        private static InfrastructureCategory GetCategory(string componentId)
+        {
+            if (componentId.StartsWith("/k8s/"))
+            {
+                return InfrastructureCategory.Kubernetes;
+            }
+            else if (componentId.StartsWith("/subscriptions/"))
+            {
+                return InfrastructureCategory.Subscription;
+            }
+
+            throw new NotSupportedException($"Not supported category for {componentId}");
+        }
+    }
+}
